+++ _CacheControl__1hour.d.ts
@@ -1,0 +1,1 @@
+export declare function _CacheControl__1hour(): any;