+++ _CacheControl__5min.d.ts
@@ -1,0 +1,1 @@
+export declare function _CacheControl__5min(): any;