+++ _ContentType__json.d.ts
@@ -1,0 +1,1 @@
+export declare function _ContentType__json(): any;