+++ _ContentType__json.ts
@@ -1,0 +1,4 @@
+import { assign } from '@ctx-core/object'
+export function _ContentType__json() {
+	return assign({ 'Content-Type': 'application/json' }, ...arguments)
+}