+++ _ContentType__svg.ts
@@ -1,0 +1,4 @@
+import { assign } from '@ctx-core/object'
+export function _ContentType__svg() {
+	return assign({ 'Content-Type': 'image/svg+xml' }, ...arguments)
+}