+++ _send_302.d.ts
@@ -1,0 +1,3 @@
+import type { Response } from 'express';
+export declare function _send_302(res: Response): void;
+export { _send_302 as _send__302 };