+++ assign_http_headers.d.ts
@@ -1,0 +1,2 @@
+export declare function assign_http_headers(ctx: any, ...header_a1: any[]): any;
+export { assign_http_headers as assign__headers__http };