+++ throw_internal_server_error.d.ts
@@ -1,0 +1,2 @@
+export declare function throw_internal_server_error(): void;
+export { throw_internal_server_error as throw__internal_server_error };